# openapi.yaml
openapi: 3.0.1
info:
  title: School Management API
  version: v1
servers:
  - url: http://localhost:8200
    description: Generated server url
paths:
...
  /integration/course/create:
    post:
      tags:
        - Course
      summary: Register new course
      operationId: createCourse
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Course'
        required: true
      responses:
        '201':
          description: Course successfully registered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
...
  /integration/student/get/name/{name}:
    get:
      tags:
        - Student
      summary: Retrieve student list that matches the given name
      operationId: getByName
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Student list obtained.
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Student'
...
  /integration/course/get/{course_id}:
    get:
      tags:
        - Course
      summary: Retrieve course
      operationId: getCourse
      parameters:
        - name: course_id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Course retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '404':
          description: Course not found.
          content:
            text/plain:
              schema:
                type: string
...
  /integration/course/get/name/{name}:
    get:
      tags:
        - Course
      summary: Retrieve course list that matches the given name
      operationId: getByName2
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Course list obtained.
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Course'
  /integration/course/get/detail/{course_id}:
    get:
      tags:
        - Course
      summary: Retrieve course detailed information
      operationId: getAllCoursesWithInfo
      parameters:
        - name: course_id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Course retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseInfo'
        '404':
          description: Course not found.
          content:
            text/plain:
              schema:
                type: string
...
  /integration/course/approve/{enrollment_id}:
    get:
      tags:
        - Course
      summary: Approve student course
      operationId: approveStudent
      parameters:
        - name: enrollment_id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Course approved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Enrollment'
        '404':
          description: Enrollment not found.
          content:
            text/plain:
              schema:
                type: string
...
  /integration/course/delete:
    delete:
      tags:
        - Course
      summary: Delete student
      operationId: deleteCourse
      parameters:
        - name: course_id
          in: query
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Course deleted.
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Course'
components:
  schemas:
    Course:
      type: object
      properties:
        id_course:
          type: integer
          format: int64
        name:
          type: string
        credits:
          type: integer
          format: int64
        spaces:
          type: integer
          format: int64
        prerequisite:
          $ref: '#/components/schemas/Course'
    Teacher:
      type: object
      properties:
        id_teacher:
          type: integer
          format: int64
        name:
          type: string
        experience_years:
          type: integer
          format: int32
        title:
          type: string
        courses:
          type: array
          items:
            $ref: '#/components/schemas/Course'
    Student:
      type: object
      properties:
        id_student:
          type: integer
          format: int64
        name:
          type: string
        credits:
          type: integer
          format: int32
        faculty:
          type: string
        semester:
          type: integer
          format: int32
        courses:
          type: array
          items:
            $ref: '#/components/schemas/Course'
...
    CourseInfo:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        credits:
          type: integer
          format: int64
        spaces:
          type: integer
          format: int64
        enrollmentCount:
          type: integer
          format: int32
        preRequisite:
          $ref: '#/components/schemas/Course'
        teachers:
          type: array
          items:
            $ref: '#/components/schemas/Teacher'
        students:
          type: array
          items:
            $ref: '#/components/schemas/Student'
    Enrollment:
      type: object
      properties:
        id_enrollment:
          type: integer
          format: int64
        id_student:
          $ref: '#/components/schemas/Student'
        id_course:
          $ref: '#/components/schemas/Course'
        approved:
          type: boolean